Introduces frontend-maven-plugin
This change makes it possible to transpile Javascript which has
important advantages.

The most important advantage is that it allows the Javascript to be
organized into ES6 modules, which to my knowledge was not possible when
using webjars.

Another advantage is that it allows modern Javascript features to be
used in development, and then transpiled into ES5 which has wider
browser support.